Move jump and spec code from unit_generic to jump_capable
